+++ frontend/src/components/widgets/Multiselect/Multiselect.tsx
@@ -20,19 +20,8 @@ import without from "lodash/without"
 import { Map as ImmutableMap } from "immutable"
 import { multiSelectOverrides } from "lib/widgetTheme"
 import { WidgetStateManager, Source } from "lib/WidgetStateManager"
-import {
-  TYPE,
-  Select as UISelect,
-  OnChangeParams,
-  StyledDropdownListItem,
-} from "baseui/select"
-import { StyledList, StyledEmptyState, OptionListProps } from "baseui/menu"
-import { FixedSizeList } from "react-window"
-
-const LIST_ITEM_HEIGHT = 40
-const EMPTY_LIST_HEIGHT = 80
-const MAX_LIST_HEIGHT = 300
-const PADDING = 16
+import { TYPE, Select as UISelect, OnChangeParams } from "baseui/select"
+import VirtualDropdown from "components/shared/VirtualDropdown"
 
 export interface Props {
   disabled: boolean
@@ -52,71 +41,6 @@ interface MultiselectOption {
   label: string
   value: string
 }
-
-interface FixedSizeListeItemProps {
-  data: { props: OptionListProps }[]
-  index: number
-  style: React.CSSProperties
-}
-
-function FixedSizeListItem(props: FixedSizeListeItemProps): ReactElement {
-  const { data, index, style } = props
-  const { item, ...restChildProps } = data[index].props
-  return (
-    <StyledDropdownListItem
-      key={item.value}
-      style={{
-        boxSizing: "border-box",
-        paddingTop: 0,
-        paddingBottom: 0,
-        display: "flex",
-        alignItems: "center",
-        ...style,
-      }}
-      {...restChildProps}
-    >
-      {item.label}
-    </StyledDropdownListItem>
-  )
-}
-
-const VirtualDropdown = React.forwardRef((props: any, ref) => {
-  const children = React.Children.toArray(
-    props.children
-  ) as React.ReactElement[]
-
-  if (!children[0] || !children[0].props.item) {
-    return (
-      <StyledList $style={{ height: `${EMPTY_LIST_HEIGHT}px` }} ref={ref}>
-        <StyledEmptyState {...children[0].props} />
-      </StyledList>
-    )
-  }
-
-  const height = Math.min(
-    MAX_LIST_HEIGHT,
-    children.length * LIST_ITEM_HEIGHT + PADDING
-  )
-
-  return (
-    <StyledList ref={ref}>
-      <FixedSizeList
-        width="100%"
-        height={height}
-        itemCount={children.length}
-        itemData={children}
-        itemKey={(index: number, data: { props: OptionListProps }[]) =>
-          data[index].props.item.value
-        }
-        itemSize={LIST_ITEM_HEIGHT}
-      >
-        {FixedSizeListItem}
-      </FixedSizeList>
-    </StyledList>
-  )
-})
-
-VirtualDropdown.displayName = "VirtualDropdown"
 
 class Multiselect extends React.PureComponent<Props, State> {
   public state: State = {
@@ -197,9 +121,10 @@ class Multiselect extends React.PureComponent<Props, State> {
           value={this.valueFromState}
           disabled={disabled}
           size={"compact"}
-          overrides={Object.assign(multiSelectOverrides, {
+          overrides={{
+            ...multiSelectOverrides,
             Dropdown: { component: VirtualDropdown },
-          })}
+          }}
         />
       </div>
     )
